Move lambda support to separate package
