+++ server/server.go
@@ -11,6 +11,7 @@ import (
 
 	"hmruntime/config"
 	"hmruntime/graphql"
+	"hmruntime/lambda"
 	"hmruntime/logger"
 
 	"github.com/rs/xid"
@@ -21,7 +22,7 @@ func Start(ctx context.Context) error {
 		Int("port", config.Port).
 		Msg("Listening for incoming requests.")
 	http.HandleFunc("/graphql", graphql.HandleGraphQLRequest)
-	http.HandleFunc("/graphql-worker", handleDgraphLambdaRequest)
+	http.HandleFunc("/graphql-worker", lambda.HandleDgraphLambdaRequest)
 	http.HandleFunc("/admin", handleAdminRequest)
 	return http.ListenAndServe(fmt.Sprintf(":%d", config.Port), nil)
 }
